hash

+++ ip-address-filters/hash_test.go
@@ -32,7 +32,7 @@ func BenchmarkHash(b *testing.B) {
 
 	b.Run("sequential int", func(b *testing.B) {
 		for i := 0; i < b.N; i++ {
-			bit := uint(i)
+			bit := uint(uint32(i))
 			if h[bit] {
 				checkMatch(uint2ip(bit))
 			}
